+++ addons/purchase/models/__init__.py
@@ -2,10 +2,11 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import account_invoice
+from . import account_tax
 from . import analytic_account
 from . import analytic_applicability
-from . import account_tax
-from . import purchase
+from . import purchase_order
+from . import purchase_order_line
 from . import product
 from . import res_company
 from . import res_config_settings
