+++ addons/purchase_stock/models/__init__.py
@@ -4,7 +4,8 @@
 from . import account_invoice
 from . import account_move_line
 from . import product
-from . import purchase
+from . import purchase_order
+from . import purchase_order_line
 from . import res_config_settings
 from . import res_partner
 from . import res_company
