+++ addons/purchase_stock/tests/test_uninstall.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 from odoo import fields
 from odoo.addons.base.models.ir_model import MODULE_UNINSTALL_FLAG
-from odoo.addons.purchase_stock.models.purchase import PurchaseOrderLine
+from odoo.addons.purchase_stock.models.purchase_order_line import PurchaseOrderLine
 from odoo.tests.common import tagged
 
 from .common import PurchaseTestCommon
